// planning-idsdp-pps-frontend/src/app/main/modules/rdpp-rtapp-management/features/dpp-tapp-list/project-details/models/similar-project-study.model.ts
import {ProjectDetailsPartB} from "./project-details-partb.model";

export class SimilarProjectStudy {
    projectConceptUuid: string;
    projectConceptMasterId: number;
    indicateIssuesMakeProject: string;
    indicateIssuesNotWork: string;
    costOfMajorItemList: any;
    itemWiseCostEstimatedList: any;
    // projectDetailsPartBRequest: ProjectDetailsPartB;
}